--- a/Manifest.toml
+++ b/Manifest.toml
@@ -28,7 +28,9 @@ version = "1.3.0"
 
 [[DataFrames]]
 deps = ["CategoricalArrays", "Compat", "DataAPI", "Future", "InvertedIndices", "IteratorInterfaceExtensions", "LinearAlgebra", "Markdown", "Missings", "PooledArrays", "Printf", "REPL", "Reexport", "SortingAlgorithms", "Statistics", "TableTraits", "Tables", "Unicode"]
-path = "/home/peterwd/Documents/Development/DataFrames"
+git-tree-sha1 = "e2d3a5ebd8c4be135b42b5d73610f9485acfe7f9"
+repo-rev = "master"
+repo-url = "https://github.com/JuliaData/DataFrames.jl.git"
 uuid = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 version = "0.21.7"