--- a/Project.toml
+++ b/Project.toml
@@ -12,11 +12,11 @@ Reexport = "0.2"
 julia = "1"
 
 [extras]
+CategoricalArrays = "324d7699-5711-5eae-9e2f-1d82baa6b597"
 Lazy = "50d2b5c4-7a5e-59d5-8109-a42b560f39c0"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-CategoricalArrays = "324d7699-5711-5eae-9e2f-1d82baa6b597"
 
 [targets]
 test = ["Lazy", "Random", "Statistics", "Test", "CategoricalArrays"]