--- a/test/dataframes.jl
+++ b/test/dataframes.jl
@@ -76,6 +76,8 @@ const ≅ = isequal
     @test @transform(df, cols("new" * "_" * "column") = :i).new_column == df.i
 
     @test @transform(df, n = 1).n == fill(1, nrow(df))
+
+    @test @transform(df, n = :i .* :g).n == [1, 2, 3, 8, 10]
 end
 
 # Defined outside of `@testset` due to use of `@eval`
@@ -190,6 +192,8 @@ end
     @test @select(df, cols(n_sym) = :i).new_column == df.i
     @test @select(df, cols(n_space) = :i)."new column" == df.i
     @test @select(df, cols("new" * "_" * "column") = :i).new_column == df.i
+
+    @test @transform(df, n = :i .* :g).n == [1, 2, 3, 8, 10]
 end
 
 # Defined outside of `@testset` due to use of `@eval`