--- a/test/performance.jl
+++ b/test/performance.jl
@@ -4,18 +4,13 @@ N = 1_000_000
 K = 10
 
 df = DataFrame(
-  id1 = rand([Symbol("id", i) for i=1:K], N),          # large groups (char)
-  id2 = rand([string("id", i) for i=1:K], N),          # large groups (char)
-  id3 = rand([string("id", i) for i=1:N÷K], N),        # small groups (char)
-  id4 = rand(1:K, N),                          # large groups (int)
-  id5 = rand(1:K, N),                          # large groups (int)
-  id6 = rand(1:N÷K, N),                        # small groups (int)
-  v1 =  rand(1:5, N),                          # int in range [1,5]
-  v2 =  rand(1:5, N),                          # int in range [1,5]
-  v3 =  rand(N)                                # numeric e.g. 23.5749
+	id = rand([Symbol("id", i) for i=1:K], N),
+	v1 = rand(1:5, N),
+	v2 = rand(1:5, N),
+	v3 = rand(N)
 );
 
-gd = groupby(df, :id1)
+gd = groupby(df, :id);
 
 function complicated_vec(x, y, z)
 	mx = mean(x)
@@ -51,7 +46,7 @@ function DataFrames_timings(df, gd)
 		:v1 => (t -> t .- mean(t)) => :res1,
 		:v2 => demean => :res2,
 		[:v1, :v2] => (+) => :res3,
-		:id4 => string => :res4,
+		:id => string => :res4,
 		[:v1, :v2, :v3] => complicated_vec => :res5,
 		[:v1, :v2, :v3] => ((a, b, c) -> @. a + b * c * c + a) => :res6a,
 		[:v1, :v2, :v3] =>
@@ -82,20 +77,15 @@ function DataFramesMeta_timings(df, gd)
 		res1 = :v1 .- mean(:v1),
 		res2 = demean(:v2),
 		res3 = :v1 + :v2,
-		res4 = string(:id4),
+		res4 = string(:id),
 		res5 = complicated_vec(:v1, :v2, :v3),
 		res6a = @.(:v1 + :v2 + :v3 * :v3 + :v1),
 		res6b = begin
-			# This zero-argument anonymous function
-			# should fix any performance costs from the
-			# @nospecialize
-			(() -> begin
-				d = Vector{Float64}(undef, length(:v1))
-				for i in eachindex(d)
-					d[i] = :v1[i] + :v2[i] * :v3[i] * :v3[i] + :v1[i]
-				end
-				d
-			end)()
+			d = Vector{Float64}(undef, length(:v1))
+			for i in eachindex(d)
+				d[i] = :v1[i] + :v2[i] * :v3[i] * :v3[i] + :v1[i]
+			end
+			d
 		end
 	)
 
@@ -115,41 +105,29 @@ println("DataFrames benchmark timings")
 println("DataFramesMeta benchmark, timings")
 @btime DataFramesMeta_timings($df, $gd);
 
-
 N = 10
 K = 10
 
 df2 = DataFrame(
-  id1 = rand([Symbol("id", i) for i=1:K], N),          # large groups (char)
-  id2 = rand([string("id", i) for i=1:K], N),          # large groups (char)
-  id3 = rand([string("id", i) for i=1:N÷K], N),        # small groups (char)
-  id4 = rand(1:K, N),                          # large groups (int)
-  id5 = rand(1:K, N),                          # large groups (int)
-  id6 = rand(1:N÷K, N),                        # small groups (int)
-  v1 =  rand(1:5, N),                          # int in range [1,5]
-  v2 =  rand(1:5, N),                          # int in range [1,5]
-  v3 =  rand(N)                                # numeric e.g. 23.5749
+	id = rand([Symbol("id", i) for i=1:K], N),
+	v1 = rand(1:5, N),
+	v2 = rand(1:5, N),
+	v3 = rand(N)
 );
 
 println("DataFramesMeta raw timing")
 @time @select(df2, res1 = :v1 .- mean(:v1));
 @time @select(df2, res2 = demean(:v2));
 @time @select(df2, res3 = :v1 + :v2);
-@time @select(df2, res4 = string(:id4));
+@time @select(df2, res4 = string(:id));
 @time @select(df2, res5 = complicated_vec(:v1, :v2, :v3));
 @time @select(df2, res6a = @.(:v1 + :v2 + :v3 * :v3 + :v1));
 @time @select(df2, res6b = begin
-			# This zero-argument anonymous function
-			# should fix any performance costs from the
-			# @nospecialize
-			(() -> begin
-				d = Vector{Float64}(undef, length(:v1))
-				for i in eachindex(d)
-					d[i] = :v1[i] + :v2[i] * :v3[i] * :v3[i] + :v1[i]
-				end
-				d
-			end)()
-		end
-	);
-nothing
+	d = Vector{Float64}(undef, length(:v1))
+	for i in eachindex(d)
+		d[i] = :v1[i] + :v2[i] * :v3[i] * :v3[i] + :v1[i]
+	end
+	d
+end);
 
+nothing