Before this PR, *not* on DataFrames master, commands take on average `.3` seconds. 

After this PR and on DataFrames master, `.08` seconds. 

Still a bit of a frustrating latency, but I think this is lots of progress and an acceptable latency!