@pdeffebach 2020-10-22
This needs benchmarking. I will revive the benchmark script in the `test` folder.