@pdeffebach 2020-10-23
Just wrote a quick performance-testing script. Things look good!

```
DataFrames benchmark timings
  193.431 ms (2001075 allocations: 240.56 MiB)
DataFramesMeta benchmark, timings
  185.151 ms (2001358 allocations: 248.20 MiB)

DataFrames raw timing
  0.210350 seconds (2.00 M allocations: 240.558 MiB, 14.87% gc time)
DataFramesMeta raw timing
  0.185518 seconds (2.00 M allocations: 240.558 MiB, 6.17% gc time)
```