@pdeffebach 2020-10-24
I have uncovered how we get a performance penalty. You get problems with 

```

function DataFrames_timings(df, gd)
	df_res = transform(df,
		:v1 => (t -> t .- mean(t)) => :res1,
		:v2 => demean => :res2,
		[:v1, :v2] => (+) => :res3,
		:id4 => string => :res4,
		[:v1, :v2, :v3] => complicated_vec => :res5,
		[:v1, :v2, :v3] => ((a, b, c) -> @. a + b * c * c + a) => :res6a,
		[:v1, :v2, :v3] =>
		(
			(a, b, c) -> begin
				d = Vector{Float64}(undef, length(a))
				for i in eachindex(d)
					d[i] = a[i] + b[i] * c[i] * c[i] + a[i]
				end
				d
			end
		) => :res6b
	)

	gd_res = combine(gd,
		:v1 => (t -> mean(t)) => :res7,
		:v2 => mean => :res8,
		[:v1, :v2] => ((t, s) -> std(t) + std(s)) => :res9,
		[:v1, :v2, :v3] => complicated_scalar => :res10,
		[:v1, :v2, :v3] => ((a, b, c)  -> first(a) + mean(b) * std(a) + last(c)) => :res11,
	)

	return(df_res, gd_res)
end

function DataFramesMeta_timings(df, gd)
	df_res = @transform(df,
		res1 = :v1 .- mean(:v1),
		res2 = demean(:v2),
		res3 = :v1 + :v2,
		res4 = string(:id4),
		res5 = complicated_vec(:v1, :v2, :v3),
		res6a = @.(:v1 + :v2 + :v3 * :v3 + :v1),
		res6b = begin
			d = Vector{Float64}(undef, length(:v1))
			for i in eachindex(d)
				d[i] = :v1[i] + :v2[i] * :v3[i] * :v3[i] + :v1[i]
			end
			return d
		end
	)


	gd_res = @based_on(gd,
		res7 = mean(:v1),
		res8 = (t -> mean(t))(:v2),
		res9 = std(:v1) + std(:v2),
		res10 = complicated_scalar(:v1, :v2, :v3),
		res11 = first(:v1) + mean(:v2) * std(:v1) + last(:v3)
	)

	return(df_res, gd_res)
end
```

`:res6b` is the problem. In DataFramesMeta, we never have a function barrier, so julia can't compile efficient code in the `for` loop. With `@nospecialize`. The options are

1. `@nospecialize`
```
            fun = quote
                (@nospecialize args...) -> begin
                    $inputargs = args
                    $body
                end
            end
```

Results: 

```
DataFrames benchmark timings
  189.276 ms (2001104 allocations: 248.19 MiB)
DataFramesMeta benchmark, timings
  419.528 ms (10999330 allocations: 393.12 MiB)
DataFrames raw timing
  0.260386 seconds (2.00 M allocations: 248.189 MiB)
DataFramesMeta raw timing
  0.592254 seconds (11.00 M allocations: 393.124 MiB, 30.86% gc time)
```

2. Just get rid of the macro 
```
            fun = quote
                (args...) -> begin
                    $inputargs = args
                    $body
                end
            end
```

Results (oddly bad)
```
DataFrames benchmark timings
  205.529 ms (2001135 allocations: 248.19 MiB)
DataFramesMeta benchmark, timings
  391.286 ms (10999330 allocations: 393.12 MiB)
DataFrames raw timing
  0.334949 seconds (2.00 M allocations: 248.189 MiB, 43.00% gc time)
DataFramesMeta raw timing
  0.476538 seconds (11.00 M allocations: 393.124 MiB, 17.27% gc time)
```

3. Define function without `vararg`
```
            fun = quote
                 $inputargs -> begin
                    $body
                end
            end
```

Results:

```
DataFrames benchmark timings
  203.278 ms (2001135 allocations: 248.19 MiB)
DataFramesMeta benchmark, timings
  203.322 ms (2001250 allocations: 255.82 MiB)
DataFrames raw timing
  0.194473 seconds (2.00 M allocations: 248.190 MiB, 7.60% gc time)
DataFramesMeta raw timing
  0.240795 seconds (2.00 M allocations: 255.822 MiB, 16.85% gc time)
```

All results are on nightly. On 1.5 there is a performance hit with `@nospecialize` that doesn't appear in 1.6. 

Is this a worthwhile tradeoff? Can we encourage users not to write loops? 

In particular, I don't think there is a way to, inside the `begin` block, specialize on the type of arguments again. Maybe one of you knows if there is an "easy fix" for users who want to opt in to performance.