@bkamins 2020-10-24
> Can we encourage users not to write loops?

I assume that with DataFramesMeta.jl we can assume users will not write loops (and can put a warning about it).
If someone wants to write a loop we should recommend - just define a function with the loop and later call it.