@pdeffebach 2020-10-24
The loop is just one example, of course. I will have to think  of other non-function operations people do. 

If there isn't anything easy to do in the body of the expression but we still want easy use, we can also set flags, like how `@byrow` will be implemented, for users to easily opt-in to faster code.