@nalimilan 2020-10-24
Yes I think we can consider that DataFramesMeta users shouldn't write loops (or they should put them in functions). As @bkamins said we can detect this and print warnings anyway. So I would drop these from the benchmarks are they probably make a big difference.