@pdeffebach 2020-10-26
I've added tests for the `:x .* :y` case, which we optimize in 1.6 but not before due to https://github.com/JuliaLang/julia/pull/37583

I also changed performance.jl to include a reasonable "fix" for writing a loop inside a `@transform` call. 

This is ready for a review / merging.