@pdeffebach 2020-10-26
I filed an issue with julia here: https://github.com/JuliaLang/julia/issues/38176

It seems like `.*` etc. is still not working. This optimization might have to wait.