@pdeffebach 2020-10-26
Okay I have purged everything related to the function re-use optimization. 

I have a strategy now to make it work but it's a lot of LOC and I would like to get a release out soon. Since the optimization relies on 1.6 behavior it doesn't make sense to push for it now. 

Ready for a review.