@nalimilan 2020-10-27
What benchmark results should I look at to see what they give with `res6b` using an anonymous function?