@pdeffebach 2020-10-27
If you mean how to configure

1. DataFrames `master`
2. Julia nightly
3. The results using `@btime`, not `@time`. 

I don't do a great job showing the improvements in compilation time but it's a bit tough to measure.