@nalimilan 2020-10-27
Sorry, I wondered whether the results you posted above were still relevant, but I got my reply. Have you run these on master on the same machine for the comparison?