@pdeffebach 2020-10-27
Okay I have updated the script to better estimate the changes in raw timings. 

Rather than estimate vs `master` I am estimating with and without `@nospecialize`. We have 

1. On nightly
2. On DataFrames master

Without `@nospecialize`

```
julia> include("test/performance.jl")
DataFrames benchmark timings
  180.904 ms (2001291 allocations: 248.20 MiB)
DataFramesMeta benchmark, timings
  196.559 ms (2001447 allocations: 255.83 MiB)
DataFramesMeta raw timing
  0.597592 seconds (523.12 k allocations: 32.467 MiB, 99.20% compilation time)
Anonymous function without `@nospecialize`
  0.056371 seconds (32.12 k allocations: 9.080 MiB, 90.52% compilation time)
Anonymous function with `@nospecialize`
  0.012096 seconds (7.74 k allocations: 8.117 MiB, 55.92% compilation time)
```

*with* `@nospecialize`

```
julia> include("test/performance.jl")
DataFrames benchmark timings
  176.683 ms (2001291 allocations: 248.20 MiB)
DataFramesMeta benchmark, timings
  185.638 ms (2001585 allocations: 255.84 MiB)
DataFramesMeta raw timing
  0.538601 seconds (523.57 k allocations: 33.295 MiB, 99.24% compilation time)
Anonymous function without `@nospecialize`
  0.055212 seconds (32.12 k allocations: 9.080 MiB, 90.50% compilation time)
Anonymous function with `@nospecialize`
  0.011967 seconds (7.75 k allocations: 8.117 MiB, 55.84% compilation time)
```

1. Looking at the difference between the "benchmark" timings --- the difference with specialize and the difference without it ---, there is a small performance hit for DataFramesMeta compared to DataFrames. Given that this decreases due to `@nospecialize`, this is probably due to improvements in compilation time, which I don't fully understand. 
2. Comparing the `DataFramesMeta raw timing` with and without `@nospecialize` we see that `@nospecialize` gets us about `.06` seconds improvement in latency. 

**However** It's really hard to pin down this number. According to the anonymous function timings we should be seeing about `.03` seconds improvement per call. But that's not quite what we have. That's actually what I get when I compare a single `@select` call to another.

Honestly it's very hard to benchmark this. I initially thought there was a bigger improvement than there seems to be after benchmarking.